add Delete&Error exception

+++ src/main/java/ru/netology/file_manager/controller/FileController.java
@@ -10,6 +10,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import ru.netology.file_manager.dto.FileListResp;
+import ru.netology.file_manager.exception.DeleteFileException;
+import ru.netology.file_manager.exception.ErrorDeleteFileException;
 import ru.netology.file_manager.exception.UploadFileException;
 import ru.netology.file_manager.model.FileInfo;
 import ru.netology.file_manager.service.FileService;
@@ -48,12 +50,15 @@ public class FileController {
 
     @CrossOrigin
     @DeleteMapping("/file")
-    public ResponseEntity<String> delete(@RequestParam("filename") String filename) throws IOException {
-        try {
+    public ResponseEntity<String> delete(@RequestParam("filename") String filename) throws IOException, DeleteFileException, ErrorDeleteFileException {
+        if (filename.isEmpty()) {
+            throw new DeleteFileException("Error input data");
+        }
+        if (fileService.checkFileExits(filename)) {
             fileService.delete(filename);
             return ResponseEntity.ok("Success deleted");
-        } catch (IllegalArgumentException e) {
-            throw new IllegalArgumentException("Error input data");
+        } else {
+            throw new ErrorDeleteFileException("Error delete file");
         }
     }
